+++ R/crew_controller_group.R
@@ -223,6 +223,47 @@ crew_class_controller_group <- R6::R6Class(
       control <- private$.select_controllers(controllers)
       all(map_lgl(control, ~.x$empty()))
     },
+    #' @description Check if the controller group is nonempty.
+    #' @details A controller is empty if it has no running tasks
+    #'   or completed tasks waiting to be retrieved with `push()`.
+    #' @return `TRUE` if the controller is empty, `FALSE` otherwise.
+    #' @param controllers Character vector of controller names.
+    #'   Set to `NULL` to select all controllers.
+    nonempty = function(controllers = NULL) {
+      control <- private$.select_controllers(controllers)
+      any(map_lgl(control, ~.x$nonempty()))
+    },
+    #' @description Number of resolved `mirai()` tasks.
+    #' @details `resolved()` is cumulative: it counts all the resolved
+    #'   tasks over the entire lifetime of the controller session.
+    #' @return Non-negative integer of length 1,
+    #'   number of resolved `mirai()` tasks.
+    #'   The return value is 0 if the condition variable does not exist
+    #'   (i.e. if the client is not running).
+    #' @param controllers Character vector of controller names.
+    #'   Set to `NULL` to select all controllers.
+    resolved = function(controllers = NULL) {
+      control <- private$.select_controllers(controllers)
+      sum(map_int(control, ~.x$resolved()))
+    },
+    #' @description Number of unresolved `mirai()` tasks.
+    #' @return Non-negative integer of length 1,
+    #'   number of unresolved `mirai()` tasks.
+    #' @param controllers Character vector of controller names.
+    #'   Set to `NULL` to select all controllers.
+    unresolved = function(controllers = NULL) {
+      control <- private$.select_controllers(controllers)
+      sum(map_int(control, ~.x$unresolved()))
+    },
+    #' @description Number of resolved `mirai()` tasks available via `pop()`.
+    #' @return Non-negative integer of length 1,
+    #'   number of resolved `mirai()` tasks available via `pop()`.
+    #' @param controllers Character vector of controller names.
+    #'   Set to `NULL` to select all controllers.
+    unpopped = function(controllers = NULL) {
+      control <- private$.select_controllers(controllers)
+      sum(map_int(control, ~.x$unpopped()))
+    },
     #' @description Check if a controller is saturated.
     #' @details A controller is saturated if the number of unresolved tasks
     #'   is greater than or equal to the maximum number of workers.
@@ -668,6 +709,80 @@ crew_class_controller_group <- R6::R6Class(
       out <- map(control, ~.x$collect(scale = scale, throttle = throttle))
       out <- tibble::new_tibble(data.table::rbindlist(out, use.names = FALSE))
       if_any(nrow(out), out, NULL)
+    },
+    #' @description Create a `promises::promise()` object to asynchronously
+    #'   pop or collect one or more tasks.
+    #' @details Please be aware that `pop()` or `collect()` will happen
+    #'   asynchronously at a some unpredictable time after the promise object
+    #'   is created, even if your local R process appears to be doing
+    #'   something completely different. This behavior is highly desirable
+    #'   in a Shiny reactive context, but please be careful as it may be
+    #'   surprising in other situations.
+    #' @return A `promises::promise()` object whose eventual value will
+    #'   be a `tibble` with results from one or more popped tasks.
+    #'   If `mode = "one"`, only one task is popped and returned (one row).
+    #'   If `mode = "all"`, then all the tasks are returned in a `tibble`
+    #'   with one row per task (or `NULL` is returned if there are no
+    #'   tasks to pop).
+    #' @param mode Character of length 1, what kind of promise to create.
+    #'   `mode` must be `"one"` or `"all"`. Details:
+    #'   * If `mode` is `"one"`, then the promise is fulfilled (or rejected)
+    #'     when at least one task is resolved and available to `pop()`.
+    #'     When that happens, `pop()` runs asynchronously, pops a result off
+    #'     the task list, and returns a value.
+    #'     If the task succeeded, then the promise
+    #'     is fulfilled and its value is the result of `pop()` (a one-row
+    #'     `tibble` with the result and metadata). If the task threw an error,
+    #'     the error message of the task is forwarded to any error callbacks
+    #'     registered with the promise.
+    #'   * If `mode` is `"all"`, then the promise is fulfilled (or rejected)
+    #'     when there are no unresolved tasks left in the controller.
+    #'     (Be careful: this condition is trivially met in the moment
+    #'     if the controller is empty and you have not submitted any tasks,
+    #'     so it is best to create this kind of promise only after you
+    #'     submit tasks.)
+    #'     When there are no unresolved tasks left,
+    #'     `collect()` runs asynchronously, pops all available results
+    #'     off the task list, and returns a value.
+    #'     If the task succeeded, then the promise
+    #'     is fulfilled and its value is the result of `collect()`
+    #'     (a `tibble` with one row per task result). If any of the tasks
+    #'     threw an error, then the first error message detected is forwarded
+    #'     to any error callbacks registered with the promise.
+    #' @param seconds_interval Positive numeric of length 1, delay in the
+    #'   `later::later()` polling interval to asynchronously check if
+    #'   the promise can be resolved.
+    #' @param scale Logical of length 1,
+    #'   whether to automatically call `scale()`
+    #'   to auto-scale workers to meet the demand of the task load.
+    #'   Scaling up on `pop()` may be important
+    #'   for transient or nearly transient workers that tend to drop off
+    #'   quickly after doing little work.
+    #'   See also the `throttle` argument.
+    #' @param throttle `TRUE` to skip auto-scaling if it already happened
+    #'   within the last `seconds_interval` seconds. `FALSE` to auto-scale
+    #'   every time `scale()` is called. Throttling avoids
+    #'   overburdening the `mirai` dispatcher and other resources.
+    #' @param controllers Not used. Included to ensure the signature is
+    #'   compatible with the analogous method of controller groups.
+    promise = function(
+      mode = "one",
+      seconds_interval = 0.1,
+      scale = TRUE,
+      throttle = TRUE,
+      controllers = NULL
+    ) {
+      # Tested in tests/interactive/test-promises.R.
+      # nocov start
+      controller_promise(
+        controller = self,
+        mode = mode,
+        seconds_interval = seconds_interval,
+        scale = scale,
+        throttle = throttle,
+        controllers = controllers
+      )
+      # nocov end
     },
     #' @description Wait for tasks.
     #' @details The `wait()` method blocks the calling R session and