+++ R/crew_package.R
@@ -21,11 +21,13 @@
 #' @importFrom cli cli_progress_bar cli_progress_done cli_progress_update
 #' @importFrom data.table rbindlist
 #' @importFrom getip getip
+#' @importFrom later later
 #' @importFrom mirai call_mirai_ daemon daemons is_mirai is_mirai_error
 #'   mirai nextget nextstream saisei status
 #' @importFrom nanonext %~>% cv cv_value mclock msleep nng_error random
 #'   .unresolved unresolved until_ tls_config
 #' @importFrom processx process
+#' @importFrom promises promise
 #' @importFrom ps ps_cmdline ps_handle ps_is_running ps_status ps_username
 #' @importFrom R6 R6Class
 #' @importFrom rlang abort as_function enquo is_installed is_named quo_squash